MF-000 Extract TravelRouteLayer from MetroMap
