+++ src/components/MetroMap.tsx
@@ -1,21 +1,12 @@
 import React, { useState, useEffect, useContext } from "react";
-import {
-  Map,
-  TileLayer,
-  Polyline,
-  FeatureGroup,
-} from "react-leaflet";
+import { Map, TileLayer } from "react-leaflet";
 import { LatLngTuple } from "leaflet";
-import { colors } from "../common/colors";
-import { Station } from "../types";
 import { DEFAULT_MAP_CENTER, DUMMY_MAP_POSITION } from "../common/mapConstants";
 import MapControl from "./map/MapControl";
-import StationMarker from "./map/StationMarker";
-import { getAllStations, getStation, getStationKeysFromTravelRoute } from "../services/util.service";
 import { TripContext } from "../contexts/TripProvider";
-import { getColorFromLineType, getInterChangeLineColor } from "../services/ui-style.service";
 import { MapContext } from "../contexts/MapProvider";
 import { MetroLineLayers } from "./map/MetroLineLayers";
+import { TravelRouteLayer } from "./map/TravelRouteLayer";
 
 export const MetroMap = () => {
   const [mapCenter, setMapCenter] = useState<LatLngTuple>(DEFAULT_MAP_CENTER);
@@ -61,67 +52,3 @@ export const MetroMap = () => {
     </div>
   );
 };
-
-const TravelRouteLayer = () => {
-  const { travelRoute } = useContext(TripContext);
-
-  const source = getStation(travelRoute.source);
-  const destination = getStation(travelRoute.destination);
-
-  if (!(source && destination)) {
-    return null;
-  }
-  const stationKeys = getStationKeysFromTravelRoute(travelRoute);
-  const allStationsInRoute = getAllStations(stationKeys);
-
-  const intermediateStations = allStationsInRoute.filter((station) =>
-    station.key !== source.key && station.key !== destination.key
-  )
-
-  return (
-    <>
-      <FeatureGroup name="travel-route">
-        {allStationsInRoute.map((currentStation, index) => {
-          if (index === 0) return null;
-          const prevStation = allStationsInRoute[index - 1];
-          const polyline = [prevStation.position, currentStation.position];
-          const color = getInterChangeLineColor(currentStation.lineType, prevStation.lineType);
-
-          return (
-            <Polyline
-              key={`travel-route-${prevStation.key}-${currentStation.key}`}
-              positions={polyline}
-              color={color}
-              weight={7}
-            />
-          );
-        })}
-      </FeatureGroup>
-      <FeatureGroup name="travel-route-station">
-        <StationMarker
-          station={source as Station}
-          fillColor={colors.sourceStation}
-          showPopup={false}
-          radius={12}
-        />
-        {intermediateStations.map((station) => {
-          return (
-            <StationMarker
-              key={`intermediate-${station.key}`}
-              station={station as Station}
-              fillColor={getColorFromLineType(station.lineType)}
-              showPopup={false}
-              radius={12}
-            />
-          );
-        })}
-        <StationMarker
-          station={destination as Station}
-          fillColor={colors.destinationStation}
-          showPopup={false}
-          radius={12}
-        />
-      </FeatureGroup>
-    </>
-  );
-}
